adding view controls for user authentication

+++ post/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect, redirect, Http404
 from .models import Post
 from .forms import PostForm
 # Create your views here.
@@ -21,31 +21,50 @@ def detailview(request, id):
 
 
 def createview(request):
+
+    if not request.user.is_authenticated():
+        return Http404()
+
     form = PostForm(request.POST or None, request.FILES or None)
+
     if form.is_valid():
         post = form.save()
         return HttpResponseRedirect(post.get_absolute_url())
+
     context = {
         'form': form
     }
+
     return render(request, 'create.html', context)
 
 
 def editview(request, id):
+    if not request.user.is_authenticated():
+        return Http404
+
     post = get_object_or_404(Post, id=id)
+
     form = PostForm(request.POST or None, request.FILES or None, instance=post)
+
     if form.is_valid():
         form.save()
         return HttpResponseRedirect(post.get_absolute_url())
+
     context = {
         'form': form
     }
+
     return render(request, 'create.html', context)
 
 
 def deleteview(request, id):
+    if not request.user.is_authenticated():
+        return Http404()
+
     post = get_object_or_404(Post, id=id)
+
     post.delete()
+
     return redirect('post:index')
